Add uploadGraph RESTful API endpoint. (check commit message for more details)
Currently the graph upload while submitting a job does not work (inside the job parameters). The backend is not yet implemented to handle uploaded graphs.
This patch adds a separate REST endpoint (uploadGraph) to upload a graph permanently to the server, just as a user would when using the web version. The user can then submit a job and use the graphID of the newly uploaded graph.
This patch has not yet been fully tested! Please check and fix bugs before merging into master.